moved overlay presenter to overlay module;
